feat(router): add option to serialize url with trailing slash
Add an option to enforce trailing slash to the URL path at the `serialize`
method of the `DefaultUrlSerializer`.
Motivation: sometimes it is desirable to always have the trailing slash
for tracking purposes. In particular, the `angular-prerender` package
reders each route into a folder with an `index.html` file, which is more
commonly served with a strailing slash. Having the option to enforce the
trailing slash would ensure the URL behaves the same either by loading
the page directly, or by navigating to the URL from another page. The
latter would go throug the Angular router module.